HT_21/requirements.txt
django==5.0
requests==2.31.0
djangorestframework==3.14.0
